+++ src/knn.c
@@ -63,7 +63,7 @@ knearneigh(int *kin, int *pnte, int *p, double *test, int *res, double *dists,
 /* Keep an extra distance if the largest current one ties with current kth */
 			if (nndist[kn] <= nndist[kinit - 1])
 			    if (++kn >= MAX_TIES - 1)
-				error("too many ties in knearneigh");
+				Rf_error("too many ties in knearneigh");
 			break;
 		    }
 	    nndist[kn] = 0.99 * DBL_MAX;