+++ src/spdep.h
@@ -1,6 +1,11 @@
-/* Copyright 2010-2021 by Roger S. Bivand. */
+/* Copyright 2010-2024 by Roger S. Bivand. */
 
-#define USE_FC_LEN_T
+#ifndef USE_FC_LEN_T
+# define USE_FC_LEN_T
+#endif
+#ifndef R_NO_REMAP
+# define R_NO_REMAP
+#endif
 #include <R.h>
 #include <Rmath.h>
 #include <Rdefines.h>